Add a mask placeholder
This is a mask that is not padded. This way we can pad gt and mask
infinitely and only request a single pixel of masked in data. This
guarantees that the center voxel of our sample will be masked
in.
This is much more stable than trying to ensure some fraction
of the full mask is masked in since many pipelines will require
a large amount of padding around the requested roi,
severly limiting the minimum size of a GT volume.
Using a separate placeholder mask from which we request
a single voxel means we can infinitely pad the GT and Mask
resulting in a more even sampling of data and the possibility
of using smaller GT crops.