Deprication warning:
If the kernel size of the task is set to 3, then this will cause errors if you are trying to use a model with input shape `(1, y, x)`.
The head should just be mapping the penultimate layer embeddings into the appropriate dimensions for the task so it desn't need a size larger than 1.